feat(bf16-vae): bf16 mixed-precision VAE training adaptation
amp_dtype flag (default fp32) + bf16-safe path across the graph-VAE stack for
the cross-node 8-card bf16 VAE run. Gated to default-fp32 so the fp32 (diffusion)
path is INTENDED bit-identical -- but NOT yet verified. Do NOT merge to main
until fp32 bit-identical is proven, to protect the running fp32 diffusion's
continued training.

- train_graph_vae.py: --amp_dtype flag + autocast wiring
- graph_salad/attention.py: bf16-safe GraphAttentionBlock (fp32-only guard rework)
- encoder.py / losses.py / motion_decoder.py / treeik_decoder.py: amp-aware
- _launch_bf16_vae_8card_xnode.sh: cross-node 8-card bf16 orchestrator
- _render_bf16_vae_recon_large.py: VAE recon QA renderer
- REPO_AUDIT_*.md / _codex_*_brief.md: bf16 audit + review docs

On bf16-vae branch (isolated worktree); does NOT touch main / noKslot_clean,
so the running fp32 diffusion is unaffected.

+++ src/models/encoder.py
@@ -84,7 +84,7 @@ class GraphAttentionBlock(nn.Module):
         mask = joint_mask.unsqueeze(1).unsqueeze(2)  # [B, 1, 1, J]
         scores = scores.masked_fill(~mask, -1e9)
 
-        attn = F.softmax(scores, dim=-1)
+        attn = F.softmax(scores.float(), dim=-1).to(scores.dtype)  # bf16-safe: fp32 softmax (fp32 path no-op)
         attn = attn.nan_to_num(0.0)  # safety: all-masked rows → 0 attention
         attn = self.dropout(attn)
 
@@ -204,7 +204,7 @@ class AnyTopGraphAttentionBlock(nn.Module):
         mask = joint_mask.unsqueeze(1).unsqueeze(2)  # [B, 1, 1, J]
         scores = scores.masked_fill(~mask, -1e9)
 
-        attn = F.softmax(scores, dim=-1)
+        attn = F.softmax(scores.float(), dim=-1).to(scores.dtype)  # bf16-safe: fp32 softmax (fp32 path no-op)
         attn = attn.nan_to_num(0.0)
         attn = self.dropout(attn)
